chore(server): add agent as valid role for client_id
