+++ common/src/testflinger_common/enums.py
@@ -102,3 +102,53 @@ class LogType(StrEnum):
 
     STANDARD_OUTPUT = "output"
     SERIAL_OUTPUT = "serial"
+
+
+class ServerRoles(StrEnum):
+    """
+    Define roles for restricted endpoints and hierarchy among them.
+
+    Implementing a custom "less-than" operator imposes an order between
+    the roles (a hierarchy) and allows for comparisons between them.
+
+    The remaining comparison operators are implemented in terms of the
+    "less-than" operator.
+
+    Note: functools.total_ordering cannot be used because `StrEnum`
+    is first derived from `str` which already provides comparison
+    operators. Thus, we need to override them explicitly.
+    """
+
+    ADMIN = "admin"
+    MANAGER = "manager"
+    CONTRIBUTOR = "contributor"
+    USER = "user"
+    AGENT = "agent"
+
+    def __lt__(self, other: "ServerRoles") -> bool:
+        """Implement of "less-than" between ServerRoles."""
+        if not isinstance(other, ServerRoles):
+            raise TypeError(
+                f"Cannot compare {type(self).__name__} "
+                f"to {type(other).__name__}"
+            )
+        _ranks = {
+            self.ADMIN: 0,
+            self.MANAGER: 1,
+            self.CONTRIBUTOR: 2,
+            self.USER: 3,
+            self.AGENT: 4,
+        }
+        return _ranks[self] > _ranks[other]
+
+    def __le__(self, other: "ServerRoles") -> bool:
+        """Implement "less-than-or-equal" between ServerRoles."""
+        return self < other or self == other
+
+    def __gt__(self, other: "ServerRoles") -> bool:
+        """Implement of "greater-than" between ServerRoles."""
+        return not (self < other or self == other)
+
+    def __ge__(self, other: "ServerRoles") -> bool:
+        """Implement of "greater-than-or-equal" between ServerRoles."""
+        return not self < other
